Http custom status codes and messages added.

+++ CatsAPI/Controllers/MyCatsController.cs
@@ -1,5 +1,6 @@
 ﻿using CatsAPI.Data;
 using CatsAPI.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -22,46 +23,69 @@ namespace CatsAPI.Controllers
 
         // GET: api/<MyCatsController>
         [HttpGet]
-        public IEnumerable<MyCat> Get()
+        public IActionResult Get()
         {
-            return _dbContext.MyCats;
+            if (_dbContext.MyCats.Any())
+                return Ok(_dbContext.MyCats);
+            else
+                return NoContent();
         }
 
         // GET api/<MyCatsController>/5
         [HttpGet("{id}")]
-        public MyCat Get(int id)
+        public IActionResult Get(int id)
         {
             var foundCat = _dbContext.MyCats.Find(id);
-            return foundCat;
+            if (foundCat is null)
+                return NotFound("This specific cat could not be found, try looking for another one.");
+            else
+                return Ok(foundCat);
         }
 
         // POST api/<MyCatsController>
         [HttpPost]
-        public void Post([FromBody] MyCat newCat)
+        public IActionResult Post([FromBody] MyCat newCat)
         {
             _dbContext.MyCats.Add(newCat);
             _dbContext.SaveChanges();
+            return StatusCode(StatusCodes.Status201Created);
         }
 
         // PUT api/<MyCatsController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] MyCat catObj)
+        public IActionResult Put(int id, [FromBody] MyCat catObj)
         {
             var foundCat = _dbContext.MyCats.Find(id);
-            foundCat.Name = catObj.Name;
-            foundCat.Type = catObj.Type;
-            foundCat.Age = catObj.Age;
-            foundCat.Weight = catObj.Weight;
-            _dbContext.SaveChanges();
+            if(foundCat is null)
+            {
+                return NotFound("This specific cat could not be found, try looking for another one.");
+            }
+            else
+            {
+                foundCat.Name = catObj.Name;
+                foundCat.Type = catObj.Type;
+                foundCat.Age = catObj.Age;
+                foundCat.Weight = catObj.Weight;
+                _dbContext.SaveChanges();
+                return Ok("Cat updated successfully.");
+            }
         }
 
         // DELETE api/<MyCatsController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             var foundCat = _dbContext.MyCats.Find(id);
-            _dbContext.MyCats.Remove(foundCat);
-            _dbContext.SaveChanges();
+            if(foundCat is null)
+            {
+                return NotFound("Could not find the cat to delete. Why are you trying to delete a cat anyway?");
+            }
+            else
+            {
+                _dbContext.MyCats.Remove(foundCat);
+                _dbContext.SaveChanges();
+                return Ok("Cat record deleted.");
+            }
         }
     }
 }
